added roles

+++ src/products/products.controller.ts
@@ -7,6 +7,11 @@ import {
   Param,
   Delete,
   UseGuards,
+  ForbiddenException,
+  ParseIntPipe,
+  Query,
+  DefaultValuePipe,
+  ParseBoolPipe,
 } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
@@ -14,7 +19,8 @@ import { UpdateProductDto } from './dto/update-product.dto';
 import { AtGuard, RolesGuard } from 'src/auth/guards';
 import { Roles } from 'src/auth/decorators';
 import { Role } from 'src/users/enums/user-role.enum';
-import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiQuery, ApiTags } from '@nestjs/swagger';
+import { UserD } from 'src/auth/decorators/users.decorator';
 
 @ApiTags('products')
 @ApiBearerAuth()
@@ -35,20 +41,45 @@ export class ProductsController {
   }
 
   @Roles(Role.ADMIN, Role.USER)
+  @ApiQuery({
+    name: 'details',
+    required: false,
+    type: 'boolean',
+    default: false,
+    description: 'Get user details with more info',
+  })
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.productsService.findOne(+id);
+  findOne(
+    @Param('id', ParseIntPipe) id: number,
+    @Query('details', new DefaultValuePipe(false), ParseBoolPipe)
+    details?: boolean,
+  ) {
+    return this.productsService.findOne(id), details;
   }
 
   @Roles(Role.ADMIN)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
+  async update(
+    @Param('id', ParseIntPipe) id: number,
+    @Body() updateProductDto: UpdateProductDto,
+    @UserD('role') token_role: Role,
+  ) {
+    if (token_role !== Role.ADMIN) {
+      throw new ForbiddenException('You are not authorized to patch producs');
+    }
     return this.productsService.update(+id, updateProductDto);
   }
 
   @Roles(Role.ADMIN, Role.USER)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(
+    @Param('id', ParseIntPipe) id: number,
+    @UserD('sub') token_id: number,
+    @UserD('role') token_role: Role,
+  ) {
+    if (token_id !== id && token_role !== Role.USER) {
+      throw new ForbiddenException('You are not authorized to patch producs');
+    }
     return this.productsService.remove(+id);
   }
 }
